# service/src/main/resources/api/service_openapi.yaml
...
openapi: 3.0.3
info:
  title: Workspace Manager API
  description: Workspace Manager REST API to manage workspaces.
  version: 0.0.1
paths:
...
  /api/workspaces/v1/{workspaceId}/resources/controlled/azure/ip:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    post:
      summary: Create a new controlled Azure IP
      operationId: createAzureIp
      tags: [ControlledAzureResource]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateControlledAzureIpRequestBody'
      responses:
        '200':
          $ref: '#/components/responses/CreateControlledAzureIpResponse'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/workspaces/v1/{workspaceId}/resources/controlled/azure/ip/{resourceId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/ResourceId'
    get:
      summary: Get a controlled Azure IP resource
      operationId: getAzureIp
      tags: [ControlledAzureResource]
      responses:
        '200':
          $ref: '#/components/responses/GetControlledAzureIpResponse'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: |
        Delete an azure IP
      operationId: deleteAzureIp
      tags: [ControlledAzureResource]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteControlledAzureIpRequest'
      responses:
        '200':
          $ref: '#/components/responses/DeleteControlledAzureIpResponse'
        '202':
          $ref: '#/components/responses/DeleteControlledAzureIpResponse'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/workspaces/v1/{workspaceId}/resources/controlled/azure/network:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    post:
      summary: Create a new controlled Azure Network
      operationId: createAzureNetwork
      tags: [ ControlledAzureResource ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateControlledAzureNetworkRequestBody'
      responses:
        '200':
          $ref: '#/components/responses/CreateControlledAzureNetworkResponse'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      description: A UUID to used to identify an object in the workspace manager
      required: true
      schema:
        type: string
        format: uuid
...
    ResourceId:
      name: resourceId
      in: path
      description: A UUID used to identify a controlled resource in a workspace.
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorReport:
      type: object
      required: [message, statusCode, causes]
      properties:
        message:
          type: string
        statusCode:
          type: integer
        causes:
          type: array
          items:
            type: string
    JobReport:
      type: object
      required: [id, status, statusCode, resultURL]
      properties:
        id:
          description: caller-provided unique identifier for the job
          type: string
        description:
          description: caller-provided description of the job
          type: string
        status:
          description: status of the job
          type: string
          enum: [RUNNING, SUCCEEDED, FAILED]
        statusCode:
          description: HTTP code providing status of the job.
          type: integer
        submitted:
          description: timestamp when the job was submitted; in ISO-8601 format
          type: string
        completed:
          description: >-
            timestamp when the job completed - in ISO-8601 format. Present if
            status is SUCCEEDED or FAILED.
          type: string
        resultURL:
          description: >-
            URL where the result of the job can be retrieved. Equivalent to a
            Location header in HTTP.
          type: string
    JobControl:
      type: object
      required: [id]
      properties:
        id:
          description: >-
            Unique identifier for the job. Best practice is for job identifier to be a UUID,
            a ShortUUID, or other globally unique identifier.
          type: string
        # TODO: In the future, notification configuration will also be part of JobControl.
    Name:
      # Note: These format restrictions are enforced by WM, not natively by
      # Swagger.
      type: string
      pattern: '^[a-zA-Z0-9][_a-zA-Z0-9]{0,62}$'
      minLength: 1
      maxLength: 63
...
    ControlledResourceIamRole:
      description: Enum containing all IAM roles on controlled resources available to users
      type: string
      enum: [ 'READER', 'WRITER', 'EDITOR']
...
    CreateControlledAzureIpRequestBody:
      description: Payload for requesting a new controlled Azure IP resource.
      type: object
      properties:
        common:
          $ref: '#/components/schemas/ControlledResourceCommonFields'
        azureIp:
          $ref: '#/components/schemas/AzureIpCreationParameters'
    CreateControlledAzureNetworkRequestBody:
      description: Payload for requesting a new controlled Azure Network resource.
      type: object
      properties:
        common:
          $ref: '#/components/schemas/ControlledResourceCommonFields'
        azureNetwork:
          $ref: '#/components/schemas/AzureNetworkCreationParameters'
...
    ControlledResourceCommonFields:
      type: object
      required: [ name, cloningInstructions, accessScope, managedBy ]
      properties:
        name:
          $ref: "#/components/schemas/Name"
        description:
          type: string
        cloningInstructions:
          $ref: '#/components/schemas/CloningInstructionsEnum'
        accessScope:
          $ref: '#/components/schemas/AccessScope'
        managedBy:
          $ref: '#/components/schemas/ManagedBy'
        privateResourceUser:
          $ref: '#/components/schemas/PrivateResourceUser'
    PrivateResourceUser:
      description: User and role information for a user of a private resource.
      type: object
      properties:
        userName:
          description: Email address for the user of the resource
          type: string
        privateResourceIamRoles:
          $ref: '#/components/schemas/PrivateResourceIamRoles'
    PrivateResourceIamRoles:
      description: >-
        List of role(s) granted to the user of a private resource. WRITER includes
        READER by definition, so if both are present only WRITER will be granted.
      type: array
      items:
        $ref: '#/components/schemas/ControlledResourceIamRole'
...
    AzureIpCreationParameters:
      description: >-
        IP-specific properties to be set on creation. These are a subset of the values
        accepted by the azure resource API
      type: object
      required: [name, region]
      properties:
        name:
          description: A valid ip name per https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules
          type: string
        region:
          description: |
            A valid region string representation per https://docs.microsoft.com/en-us/dotnet/api/microsoft.azure.management.resourcemanager.fluent.core.region.
            It is best to look at the Region class definition for valid strings (e.g. 'eastus' == Region.US_EAST.toString())
          type: string
    AzureNetworkCreationParameters:
      description: >-
        Network-specific properties to be set on creation. These are a subset of the values
        accepted by the azure resource API
      type: object
      required: [ name, region ] # TODO
      properties:
        name:
          description: A valid Network name per https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules
          type: string
        region:
          description: |
            A valid region string representation per https://docs.microsoft.com/en-us/dotnet/api/microsoft.azure.management.resourcemanager.fluent.core.region.
            It is best to look at the Region class definition for valid strings (e.g. 'eastus' == Region.US_EAST.toString())
          type: string
    CreatedControlledAzureIp:
      description: Response payload for requesting a new azure IP
      type: object
      properties:
        resourceId:
          description: UUID of a newly-created resource.
          type: string
          format: uuid
        azureIp:
          $ref: '#/components/schemas/AzureIpResource'
    CreatedControlledAzureNetwork:
      description: Response payload for requesting a new azure network
      type: object
      properties:
        resourceId:
          description: UUID of a newly-created resource.
          type: string
          format: uuid
        azureNetwork:
          $ref: '#/components/schemas/AzureNetworkResource'
...
    DeleteControlledAzureIpRequest:
      type: object
      required: [ jobControl ]
      properties:
        jobControl:
          $ref: '#/components/schemas/JobControl'
    DeleteControlledAzureIpResult:
      type: object
      properties:
        jobReport:
          $ref: '#/components/schemas/JobReport'
        errorReport:
          $ref: '#/components/schemas/ErrorReport'
...
    ControlledResourceMetadata:
      type: object
      properties:
        accessScope:
          $ref: '#/components/schemas/AccessScope'
        managedBy:
          $ref: '#/components/schemas/ManagedBy'
        privateResourceUser:
          $ref: '#/components/schemas/PrivateResourceUser'
...
    AzureIpAttributes:
      description: >-
        IP properties included in post-creation get and update. Others must be retrieved from Azure using the name.
      type: object
      properties:
        ipName:
          description: Name of created ip (not the resource name).
          type: string
        region:
          description: Azure region of created ip
          type: string
    AzureNetworkAttributes:
      description: >-
        Network properties included in post-creation get and update. Others must be retrieved from Azure using the name.
      type: object
      properties:
        networkName:
          description: Name of created network (not the resource name).
          type: string
        region:
          description: Azure region of created Network
          type: string
...
    AzureIpResource:
      type: object
      description: Description of an Azure Ip
      required: [metadata, attributes]
      properties:
        metadata:
          description: the resource metadata common to all resources
          $ref: '#/components/schemas/ResourceMetadata'
        attributes:
          $ref: '#/components/schemas/AzureIpAttributes'
    AzureNetworkResource:
      type: object
      description: Description of an Azure Network
      required: [ metadata, attributes ]
      properties:
        metadata:
          description: the resource metadata common to all resources
          $ref: '#/components/schemas/ResourceMetadata'
        attributes:
          $ref: '#/components/schemas/AzureNetworkAttributes'
...
    ResourceMetadata:
      type: object
      properties:
        workspaceId:
          type: string
          format: uuid
        resourceId:
          type: string
          format: uuid
        name:
          $ref: "#/components/schemas/Name"
        description:
          type: string
        resourceType:
          $ref: '#/components/schemas/ResourceType'
        stewardshipType:
          $ref: '#/components/schemas/StewardshipType'
        cloudPlatform:
          $ref: '#/components/schemas/CloudPlatform'
        cloningInstructions:
          $ref: '#/components/schemas/CloningInstructionsEnum'
        controlledResourceMetadata:
          description: Present if stewardship type is CONTROLLED
          $ref: '#/components/schemas/ControlledResourceMetadata'
...
    AccessScope:
      type: string
      description: Specifies the resource as shared or private
      enum: ['SHARED_ACCESS', 'PRIVATE_ACCESS']
    CloudPlatform:
      type: string
      description: Enum representing a cloud platform type.
      enum: ['AZURE', 'GCP']
    CloningInstructionsEnum:
      type: string
      description: Instructions for copying this reference when cloning the workspace
      enum: ['COPY_NOTHING', 'COPY_DEFINITION', 'COPY_RESOURCE', 'COPY_REFERENCE']
    ManagedBy:
      type: string
      description: Specifies the controller of the resource, workspace users or an application
      enum: [ 'USER', 'APPLICATION' ]

    # The resource types are common to both stewardship types
    ResourceType:
      description: Enum containing valid resource types. Used for enumeration
      type: string
      enum:
        - AI_NOTEBOOK
        - BIG_QUERY_DATASET
        - DATA_REPO_SNAPSHOT
        - GCS_BUCKET
        - AZURE_IP
        - AZURE_NETWORK
    StewardshipType:
      description: Enum containing valid stewardship types. Used for enumeration
      type: string
      enum: ['REFERENCED', 'CONTROLLED']
  responses:
...
    CreateControlledAzureIpResponse:
      description: Response to create controlled azure ip
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreatedControlledAzureIp'
    DeleteControlledAzureIpResponse:
      description: Response Payload for deleting an azure ip
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DeleteControlledAzureIpResult'
    GetControlledAzureIpResponse:
      description: Response to get ip
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AzureIpResource'
    CreateControlledAzureNetworkResponse:
      description: Response to create controlled azure network
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreatedControlledAzureNetwork'
...
    PermissionDenied:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReport'
    NotFound:
      description: Not found (or unauthorized)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReport'
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReport'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    authorization:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization
    ## for AzurePoC - we do not do authn. Just accept the user email and objectId as user/pwd
    basicAuth:
      type: http
      scheme: basic
security:
  - bearerAuth: []
  - authorization: [openid, email, profile]
  - basicAuth: []